refactor: move scan tests
- move scanner tests to test_scanner
- merge test_commit in test_scannable
